+++ app/Plugin/Init.php
@@ -101,6 +101,9 @@ class Init {
 		// init support for command palette.
 		Commands::get_instance()->init();
 
+		// init the diagnose tool.
+		Diagnostics::get_instance()->init();
+
 		// install db tables on plugin-installation.
 		add_action( 'personio_integration_install_db_tables', array( $this, 'install_db_tables' ) );